+++ src/main/kotlin/shop/itbug/fluttercheckversionx/actions/AssetsAutoGenerateClassAction.kt
@@ -11,13 +11,13 @@ import shop.itbug.fluttercheckversionx.util.MyDartPsiElementUtil
 
 /// 自动正常资产文件调用
 ///
-class AssetsAutoGenerateClassAction : MyAction(PluginBundle.getLazyMessage("assets.gen")) {
+class AssetsAutoGenerateClassAction : MyAction() {
 
     override fun actionPerformed(e: AnActionEvent) {
         val project = e.getData(CommonDataKeys.PROJECT)!!
         val vf = e.getData(CommonDataKeys.VIRTUAL_FILE)!!
         val name = vf.name
-        val userSetting = GenerateAssetsClassConfig.getGenerateAssetsSetting()
+        val userSetting = GenerateAssetsClassConfig.getGenerateAssetsSetting(project)
         val isOk =
             if (userSetting.dontTip) true else AssetsAutoGenerateClassActionConfigDialog(project).showAndGet()
         if (!isOk) {
@@ -29,7 +29,7 @@ class AssetsAutoGenerateClassAction : MyAction(PluginBundle.getLazyMessage("asse
     override fun update(e: AnActionEvent) {
         val vf = e.getData(CommonDataKeys.VIRTUAL_FILE)
         e.presentation.isEnabledAndVisible = vf != null && vf.isDirectory && e.project != null
-
+        e.presentation.text = PluginBundle.get("assets.gen")
         super.update(e)
     }