+++ src/main/kotlin/shop/itbug/fluttercheckversionx/dialog/JsonToDartBaseConfig.kt
@@ -49,10 +49,9 @@ fun Panel.nameRuleConfig(onChange: NameRuleConfig, init: OnInit? = null): Collap
 
 
 // freezed新版本的设置
-
-
 data class SaveToDirectoryModelOnChange(
-    val onDirectoryChange: KMutableProperty0<String>,
+    val directoryGet: () -> String,
+    val directorySet: (string: String) -> Unit,
     val onFilenameChange: KMutableProperty0<String>,
     val onOpenInEditor: KMutableProperty0<Boolean>,
 )
@@ -76,7 +75,7 @@ fun Row.saveToDirectoryConfig(
                     }
             }
             row(PluginBundle.get("g.3")) {
-                MyRowBuild.folder(this, onChange.onDirectoryChange, project)
+                MyRowBuild.folder(this, onChange.directoryGet, onChange.directorySet, project)
             }
             row {
                 checkBox(PluginBundle.get("freezed.gen.base.open.in.editor")).bindSelected(onChange.onOpenInEditor)
@@ -94,19 +93,37 @@ object FreezedNewSetting {
             cell(box).bindItem({ config.freezedVersion }, {
                 config.freezedVersion = it ?: FreezedVersion.DefaultVersion
             })
-            cell(box2).bindItem({ config.freezedClassType },
+            cell(box2).bindItem(
+                { config.freezedClassType },
                 { config.freezedClassType = it ?: FreezedClassType.Sealed })
         }
     }
 }
 
 
-private object MyRowBuild {
+object MyRowBuild {
     /// 2024.3
+    fun folder(row: Row, getter: () -> String, setter: (String) -> Unit, project: Project) {
+        row.textFieldWithBrowseButton(
+            FileChooserDescriptorFactory.createSingleFolderDescriptor(),
+            project,
+        ).bindText(getter, setter).align(Align.FILL).addValidationRule(VerifyFileDir.ERROR_MSG) {
+            VerifyFileDir.validDirByComponent(it)
+        }.validationOnInput {
+            if (VerifyFileDir.validDirByComponent(it)) {
+                return@validationOnInput ValidationInfoBuilder(it.textField).error(VerifyFileDir.ERROR_MSG)
+            }
+            return@validationOnInput null
+        }
+    }
+
+    /// 2023.2
 //    fun folder(row: Row, onChange: KMutableProperty0<String>, project: Project) {
 //        row.textFieldWithBrowseButton(
-//            FileChooserDescriptorFactory.createSingleFolderDescriptor(),
+//            PluginBundle.get("select_a_folder"),
 //            project,
+//            FileChooserDescriptorFactory.createSingleFolderDescriptor(),
+//            { it.path },
 //        ).bindText(onChange).align(Align.FILL).addValidationRule(VerifyFileDir.ERROR_MSG) {
 //            VerifyFileDir.validDirByComponent(it)
 //        }.validationOnInput {
@@ -116,23 +133,6 @@ private object MyRowBuild {
 //            return@validationOnInput null
 //        }
 //    }
-
-    /// 2023.2
-    fun folder(row: Row, onChange: KMutableProperty0<String>, project: Project) {
-        row.textFieldWithBrowseButton(
-            PluginBundle.get("select_a_folder"),
-            project,
-            FileChooserDescriptorFactory.createSingleFolderDescriptor(),
-            { it.path },
-        ).bindText(onChange).align(Align.FILL).addValidationRule(VerifyFileDir.ERROR_MSG) {
-            VerifyFileDir.validDirByComponent(it)
-        }.validationOnInput {
-            if (VerifyFileDir.validDirByComponent(it)) {
-                return@validationOnInput ValidationInfoBuilder(it.textField).error(VerifyFileDir.ERROR_MSG)
-            }
-            return@validationOnInput null
-        }
-    }
 }
 
 enum class NameFormatRule(val format: CaseFormat, val eg: String) {