+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## [5.3.1] - 2025-03-03
+
+* Fix: dart function comment generator tool error (dart comment)
+* Fix: dart assets hover document preview bug (with image) (dart assets preview)
+* New: Supports the operation of clicking on the asset path to quickly open files（Need to hold ctrl）
+
 ## [5.3.0] - 2025-02-28
 
 * New features: Added assets asset path detection mark (dart)